+++ src/codegen/templates/cpp/ramble.hpp.hbs
