+++ src/lib.rs
@@ -2,7 +2,7 @@ pub use packet::RambleConfig;
 pub use parse::Scanner;
 pub use util::load_ramble_file;
 
+pub mod codegen;
 mod packet;
 mod parse;
-pub mod targets;
 mod util;
